refactor: reorganize charge sensor utilities and enhance peak stability analysis
- Moved utility functions related to charge sensor routines into a new `utils` module for better organization and modularity.
- Introduced new functions for peak stability analysis, including `calculate_local_slope`, `calculate_voltage_noise`, and `measure_peak_stability`, to improve the robustness of peak selection.
- Updated existing routines to utilize the new utility functions, streamlining the codebase and enhancing maintainability.
- Enhanced documentation for new and existing functions to provide clearer insights into their functionality and usage.
- Improved test coverage for the new utility functions and stability analysis, ensuring comprehensive validation of charge sensor performance.
