+++ stanza/routines/builtins/utils/peak_fitting.py
@@ -12,6 +12,9 @@ except ImportError:
 
 from dataclasses import dataclass
 
+from stanza.exceptions import RoutineError
+from stanza.logger.session import LoggerSession
+
 logger = getLogger(__name__)
 
 
@@ -641,3 +644,311 @@ def fit_peak_multi_model(
     )
 
     return fitted_peak
+
+
+def fit_and_log_single_peak(
+    peak_idx: int,
+    peak_number: int,
+    start_idx: int,
+    end_idx: int,
+    aggregated_currents: np.ndarray,
+    aggregated_voltages: np.ndarray,
+    analysis_session: LoggerSession | None,
+) -> FittedPeak | None:
+    """
+    Fit a single peak and log its analysis results.
+
+    Args:
+        peak_idx: Index of peak in aggregated trace
+        peak_number: Sequential number of peak (1-based, for logging)
+        start_idx: Start index of window in aggregated trace
+        end_idx: End index of window in aggregated trace
+        aggregated_currents: Full aggregated current trace
+        aggregated_voltages: Full aggregated voltage trace
+        analysis_session: Logging session for saving analysis results
+
+    Returns:
+        FittedPeak object if successful, None otherwise
+    """
+    window_currents = aggregated_currents[start_idx:end_idx]
+    window_indices = np.arange(len(window_currents))
+    peak_idx_in_window = peak_idx - start_idx
+
+    if peak_idx_in_window < 0 or peak_idx_in_window >= len(window_currents):
+        return None
+
+    try:
+        fitted_peak = fit_peak_multi_model(
+            window_currents=window_currents,
+            window_indices=window_indices,
+            peak_idx_in_window=peak_idx_in_window,
+            aggregated_voltages=aggregated_voltages,
+            window_start_idx=start_idx,
+            window_end_idx=end_idx,
+            peak_idx_aggregated=peak_idx,
+        )
+
+        if analysis_session:
+            analysis_session.log_analysis(
+                name=f"peak_{peak_number}_multi_model_fit",
+                data={
+                    "peak_number": peak_number,
+                    "peak_idx": fitted_peak.peak_idx,
+                    "best_model": fitted_peak.best_model,
+                    "quality_score": fitted_peak.quality_score,
+                    "sensitivity": fitted_peak.sensitivity,
+                    "sensitivity_voltage": fitted_peak.sensitivity_voltage,
+                    "peak_voltage": fitted_peak.peak_voltage,
+                    "lorentzian": {
+                        "amplitude": fitted_peak.lorentzian_fit.amplitude,
+                        "center_idx": fitted_peak.lorentzian_fit.center_idx,
+                        "width": fitted_peak.lorentzian_fit.width,
+                        "offset": fitted_peak.lorentzian_fit.offset,
+                        "r_squared": fitted_peak.lorentzian_fit.r_squared,
+                        "rmse": fitted_peak.lorentzian_fit.rmse,
+                        "aicc": fitted_peak.lorentzian_fit.aicc,
+                        "fwhm": fitted_peak.lorentzian_fit.fwhm,
+                        "area": fitted_peak.lorentzian_fit.area,
+                        "skew_resid": fitted_peak.lorentzian_fit.skew_resid,
+                    },
+                    "sech2": {
+                        "amplitude": fitted_peak.sech2_fit.amplitude,
+                        "center_idx": fitted_peak.sech2_fit.center_idx,
+                        "width": fitted_peak.sech2_fit.width,
+                        "offset": fitted_peak.sech2_fit.offset,
+                        "r_squared": fitted_peak.sech2_fit.r_squared,
+                        "rmse": fitted_peak.sech2_fit.rmse,
+                        "aicc": fitted_peak.sech2_fit.aicc,
+                        "fwhm": fitted_peak.sech2_fit.fwhm,
+                        "area": fitted_peak.sech2_fit.area,
+                        "skew_resid": fitted_peak.sech2_fit.skew_resid,
+                    },
+                    "voigt": {
+                        "amplitude": fitted_peak.voigt_fit.amplitude,
+                        "center_idx": fitted_peak.voigt_fit.center_idx,
+                        "width": fitted_peak.voigt_fit.width,
+                        "offset": fitted_peak.voigt_fit.offset,
+                        "eta": fitted_peak.voigt_fit.eta,
+                        "r_squared": fitted_peak.voigt_fit.r_squared,
+                        "rmse": fitted_peak.voigt_fit.rmse,
+                        "aicc": fitted_peak.voigt_fit.aicc,
+                        "fwhm": fitted_peak.voigt_fit.fwhm,
+                        "area": fitted_peak.voigt_fit.area,
+                        "skew_resid": fitted_peak.voigt_fit.skew_resid,
+                    },
+                },
+            )
+
+        return fitted_peak
+
+    except Exception as e:
+        logger.warning("Failed to fit peak at index %s: %s", peak_idx, e)
+        return None
+
+
+def calculate_peak_window_bounds(
+    peak_idx: int,
+    peak_index: int,
+    peak_indices: list[int],
+    trace_length: int,
+    window_fraction: float = 0.8,
+    default_window_half_width: int = 128,
+) -> tuple[int, int] | None:
+    """
+    Calculate window boundaries for fitting a peak based on its position.
+
+    Uses spacing between peaks to determine appropriate window size, handling
+    edge cases for first, middle, and last peaks.
+
+    Args:
+        peak_idx: Index of the current peak in the aggregated trace
+        peak_index: Position of peak in the sorted peak_indices list (0-based)
+        peak_indices: Sorted list of all peak indices
+        trace_length: Total length of the aggregated trace
+        window_fraction: Fraction of inter-peak distance to use for window (default: 0.8)
+        default_window_half_width: Default half-width in points when peaks are widely spaced (default: 128)
+
+    Returns:
+        Tuple of (start_idx, end_idx) for the window, or None if invalid
+    """
+    if len(peak_indices) == 1:
+        start_idx = 0
+        end_idx = trace_length
+    elif peak_index == 0:
+        distance_to_next = peak_indices[peak_index + 1] - peak_idx
+        distance_to_start = peak_idx
+
+        start_idx = (
+            int(peak_idx - window_fraction * distance_to_start)
+            if distance_to_start < default_window_half_width
+            else peak_idx - default_window_half_width
+        )
+        end_idx = (
+            int(peak_idx + window_fraction * distance_to_next)
+            if distance_to_next < default_window_half_width
+            else peak_idx + default_window_half_width
+        )
+    elif peak_index == len(peak_indices) - 1:
+        distance_to_end = trace_length - peak_idx
+        distance_to_previous = peak_idx - peak_indices[peak_index - 1]
+
+        start_idx = (
+            int(peak_idx - distance_to_previous * window_fraction)
+            if distance_to_previous < default_window_half_width
+            else peak_idx - default_window_half_width
+        )
+        end_idx = (
+            int(peak_idx + distance_to_end * window_fraction)
+            if distance_to_end < default_window_half_width
+            else peak_idx + default_window_half_width
+        )
+    else:
+        distance_to_next = peak_indices[peak_index + 1] - peak_idx
+        distance_to_previous = peak_idx - peak_indices[peak_index - 1]
+
+        start_idx = (
+            int(peak_idx - window_fraction * distance_to_previous)
+            if distance_to_previous < default_window_half_width
+            else peak_idx - default_window_half_width
+        )
+        end_idx = (
+            int(peak_idx + window_fraction * distance_to_next)
+            if distance_to_next < default_window_half_width
+            else peak_idx + default_window_half_width
+        )
+
+    start_idx = int(max(0, start_idx))
+    end_idx = int(min(trace_length, end_idx))
+
+    if start_idx >= peak_idx:
+        start_idx = int(max(0, peak_idx - 5))
+    if end_idx <= peak_idx:
+        end_idx = int(min(trace_length, peak_idx + 5))
+
+    if peak_idx < start_idx or peak_idx >= end_idx:
+        return None
+
+    return (start_idx, end_idx)
+
+
+def normalize_sensitivity_scores(fitted_peaks: list[FittedPeak]) -> None:
+    """
+    Normalize sensitivity scores across all fitted peaks using min-max scaling.
+
+    Args:
+        fitted_peaks: List of FittedPeak objects to normalize
+    """
+    if not fitted_peaks:
+        return
+
+    sensitivities = [peak.sensitivity for peak in fitted_peaks]
+    min_sens = min(sensitivities)
+    max_sens = max(sensitivities)
+    sens_range = max_sens - min_sens
+
+    if sens_range > 0:
+        for peak in fitted_peaks:
+            peak.sensitivity_score = (peak.sensitivity - min_sens) / sens_range
+    else:
+        for peak in fitted_peaks:
+            peak.sensitivity_score = 1.0
+
+
+def calculate_quality_scores(fitted_peaks: list[FittedPeak]) -> None:
+    """
+    Calculate quality scores for all fitted peaks using normalized sensitivity scores.
+
+    Args:
+        fitted_peaks: List of FittedPeak objects (must have sensitivity_score set)
+    """
+    for peak in fitted_peaks:
+        best_fit = getattr(peak, f"{peak.best_model.lower()}_fit")
+
+        y_max = (
+            float(np.max(np.abs(peak.window_currents)))
+            if len(peak.window_currents) > 0
+            else 1.0
+        )
+
+        if peak.sensitivity_score is None:
+            raise RoutineError(
+                f"Sensitivity score not set for peak at index {peak.peak_idx}. "
+                "Ensure normalize_sensitivity_scores() is called first."
+            )
+        peak.quality_score = calculate_quality_score(
+            r_squared=best_fit.r_squared,
+            rmse=best_fit.rmse,
+            y_max=y_max,
+            skew=best_fit.skew_resid,
+            sensitivity_score=peak.sensitivity_score,
+        )
+
+
+def analyze_find_first_peak_voltages(
+    aggregated_currents: np.ndarray,
+    aggregated_voltages: np.ndarray,
+    peak_indices_aggregated: list[int],
+    analysis_session: LoggerSession | None,
+    window_fraction: float = 0.8,
+    default_window_half_width: int = 128,
+) -> list[FittedPeak]:
+    """
+    Analyze and fit multiple peaks in aggregated current trace.
+
+    Args:
+        aggregated_currents: Aggregated current trace data
+        aggregated_voltages: Aggregated voltage trace data
+        peak_indices_aggregated: List of peak indices in aggregated trace
+        analysis_session: Logging session for saving analysis results
+        window_fraction: Fraction of inter-peak distance to use for window (default: 0.8)
+        default_window_half_width: Default half-width in points when peaks are widely spaced (default: 128)
+
+    Returns:
+        List of FittedPeak objects with quality metrics
+    """
+    fitted_peaks = []
+    peak_indices_aggregated = sorted(set(peak_indices_aggregated))
+
+    for i, peak_idx in enumerate(peak_indices_aggregated):
+        window_bounds = calculate_peak_window_bounds(
+            peak_idx=peak_idx,
+            peak_index=i,
+            peak_indices=peak_indices_aggregated,
+            trace_length=len(aggregated_currents),
+            window_fraction=window_fraction,
+            default_window_half_width=default_window_half_width,
+        )
+
+        if window_bounds is None:
+            continue
+
+        start_idx, end_idx = window_bounds
+
+        fitted_peak = fit_and_log_single_peak(
+            peak_idx=peak_idx,
+            peak_number=i + 1,
+            start_idx=start_idx,
+            end_idx=end_idx,
+            aggregated_currents=aggregated_currents,
+            aggregated_voltages=aggregated_voltages,
+            analysis_session=analysis_session,
+        )
+
+        if fitted_peak is not None:
+            fitted_peaks.append(fitted_peak)
+
+    if analysis_session:
+        analysis_session.log_analysis(
+            name="peak_fitting_summary",
+            data={
+                "num_peaks_detected": len(peak_indices_aggregated),
+                "num_peaks_fitted": len(fitted_peaks),
+                "fitted_peak_indices": [int(p.peak_idx) for p in fitted_peaks],
+            },
+        )
+
+    if fitted_peaks:
+        normalize_sensitivity_scores(fitted_peaks)
+        calculate_quality_scores(fitted_peaks)
+
+    return fitted_peaks
